+++ django_cassandra_engine/connection.py
@@ -9,6 +9,11 @@ class CassandraConnection(object):
         self.hosts = options.get('HOST').split(',')
         self.keyspace = options.get('NAME')
         self.options = options.get('OPTIONS', {})
+        self.connection_options = self.options.get('connection', {})
+        self.connection_options['lazy_connect'] = \
+            self.connection_options.get('lazy_connect', True)
+        self.connection_options['retry_connect'] = \
+            self.connection_options.get('retry_connect', True)
         self.setup()
 
     def setup(self):
@@ -16,11 +21,7 @@ class CassandraConnection(object):
             # already connected
             return
 
-        connection.setup(
-            self.hosts,
-            self.keyspace,
-            **self.options.get('connection', {})
-        )
+        connection.setup(self.hosts, self.keyspace, **self.connection_options)
 
     def commit(self):
         pass